refactor: rearchitecture infra layer
