+++ src/main/kotlin/com/kroffle/knitting/infra/calculator/data/CalculateType.kt
@@ -1,4 +1,4 @@
-package com.kroffle.knitting.infra.design.calculator.data
+package com.kroffle.knitting.infra.calculator.data
 
 enum class CalculateType(
     val gaugeType: GaugeType,
