+++ src/main/kotlin/com/kroffle/knitting/infra/calculator/data/DraftInlineStyleRange.kt
@@ -1,4 +1,4 @@
-package com.kroffle.knitting.infra.design.calculator.data
+package com.kroffle.knitting.infra.calculator.data
 
 import com.fasterxml.jackson.annotation.JsonIgnore
