--- src/main/kotlin/com/kroffle/knitting/infra/design/calculator/data/GaugeType.kt
@@ -1,6 +1,0 @@
-package com.kroffle.knitting.infra.design.calculator.data
-
-enum class GaugeType {
-    Stitch,
-    Row
-}
