+++ src/main/kotlin/com/kroffle/knitting/infra/persistence/design/DBDesignRepository.kt
@@ -1,6 +1,6 @@
-package com.kroffle.knitting.infra.design
+package com.kroffle.knitting.infra.persistence.design
 
-import com.kroffle.knitting.infra.design.entity.DesignEntity
+import com.kroffle.knitting.infra.persistence.design.entity.DesignEntity
 import org.springframework.data.repository.reactive.ReactiveCrudRepository
 import reactor.core.publisher.Flux
