+++ src/main/kotlin/com/kroffle/knitting/infra/persistence/knitter/DBKnitterRepository.kt
@@ -1,6 +1,6 @@
-package com.kroffle.knitting.infra.knitter
+package com.kroffle.knitting.infra.persistence.knitter
 
-import com.kroffle.knitting.infra.knitter.entity.KnitterEntity
+import com.kroffle.knitting.infra.persistence.knitter.entity.KnitterEntity
 import org.springframework.data.repository.reactive.ReactiveCrudRepository
 import reactor.core.publisher.Mono
